packages/core_py/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agt-publisher-core"
version = "0.1.0"
description = "Shared core for WordPress publishing + SEO ops (Camp Lakota system template)"
requires-python = ">=3.8"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "Pillow>=10.0.0",
  "jsonschema>=4.17.0",
  "rich>=13.0.0",
  "python-dateutil>=2.8.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"agt_publisher_core" = ["schemas/*.json"]

[tool.setuptools.packages.find]
where = ["src"]
